Clean up window management code a bit
